Stall interface vibecoded :)

+++ src/main/java/de/infokurs/Info_Projekt_12_2_2026/model/unicorns/FireUnicornExample.java
@@ -16,7 +16,7 @@ public class FireUnicornExample extends Unicorn{
     }
 
     @Override
-    public double getRps() {
+    public int getRps() {
         return 12345;
     }
